fix NSInteger format warning

+++ SVPullToRefresh/UIScrollView+SVInfiniteScrolling.m
@@ -242,8 +242,8 @@ UIEdgeInsets scrollViewOriginalContentInsets;
     if(!_loadingImageView) {
         _loadingImageView = [[UIImageView alloc] initWithFrame:CGRectMake(0, 0, 32, 32)];
         NSMutableArray *animationImages = [[NSMutableArray alloc] initWithCapacity:12];
-        for (NSInteger index = 0; index <= 12; index++) {
-            NSString *imageName = [NSString stringWithFormat:@"loading_%li", index];
+        for (int index = 0; index <= 12; index++) {
+            NSString *imageName = [NSString stringWithFormat:@"loading_%d", index];
             UIImage *image = [UIImage imageNamed:imageName inBundle:[NSBundle bundleForClass:self.class] compatibleWithTraitCollection:nil];
             [animationImages addObject:image];
         }
